requirements.txt
gradio==5.44.1
openai==1.102.0
python-dotenv==1.1.1
